make the docker process run as user 999
It also exits with a helpful error message if the state directory isn’t writable
or accessible by the uid 999.

+++ scripts/launch/connect-to-cluster/default.nix
@@ -75,6 +75,13 @@ let
 in pkgs.writeScript "${executable}-connect-to-${environment}" ''
   #!${pkgs.stdenv.shell}
 
+  mkdir -p ${stateDir}
+
+  if [[ ! -w ${stateDir} || ! -x ${stateDir} ]]; then
+    echo "The state dir (${stateDir}) must be owned by user id $(id -u), exiting" 2>&1
+    exit 1
+  fi
+
   if [[ "$1" == "--delete-state" ]]; then
     echo "Deleting ${stateDir} ... "
     rm -Rf ${stateDir}
